Update BookTable genre to be hardcoded
There was an issue where filtering, then searching the list would result
in a smaller genre list than expected as the genre list is created from
the fetched book data.

To fix this I've updated the initial genre array to be a hardcoded list
as the genres they don't change.